+++ lib/poisson_disc.dart
@@ -1,7 +1,0 @@
-library poisson_disc;
-
-import 'dart:collection';
-import 'dart:math';
-import 'package:vector_math/vector_math.dart';
-
-part 'src/bridson2.dart';